main_service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
google-genai>=1.14.0
pydantic==2.5.3
python-multipart==0.0.6
aiofiles==23.2.1
google-cloud-texttospeech>=2.17.0
numpy
Pillow
